finish get_nodes_for_embeddings
